(fix) add safe accessor on `name` (#37)
In the case where the select has no value, notion will return a nil
value on the select property. This was assumed to be present, so calling
`.name` on nil returned an exception.

This adds the safe accessor in the event it's not set, the returned
property value from `NotionToMd::Page::CustomProperty` will also be nil.